Backend/requirements.txt
asn1crypto==1.5.1
black==23.11.0
blinker==1.9.0
click==8.1.8
flake8==6.1.0
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.20.1
mccabe==0.7.0
mypy==1.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pg8000==1.29.4
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.3
pytest-mock==3.14.0
pytest-timeout==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.1
scramp==1.4.5
six==1.17.0
SQLAlchemy==2.0.25
typing_extensions==4.12.2
Werkzeug==3.1.3
fastapi==0.95.0
uvicorn==0.20.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.24.1
python-multipart==0.0.6
bcrypt==4.0.1
slowapi==0.1.8
alembic[tz]

